Status check openNewTabTest UI test
