+++ automation/taskcluster/androidTest/flank-arm64-v8a.yml
@@ -20,7 +20,7 @@ gcloud:
   performance-metrics: true
 
   test-targets:
-    - package org.mozilla.reference.browser.ui
+    - class org.mozilla.reference.browser.ui.ContextMenusTest#openNewTabTest
 
   device:
     - model: Pixel2.arm
@@ -30,6 +30,6 @@ gcloud:
 flank:
   project: GOOGLE_PROJECT
   max-test-shards: 50
-  num-test-runs: 1
+  num-test-runs: 5
   output-style: compact
   full-junit-result: true
